# fix: speed up labelsToString