@ellistarn 2023-04-17
I know @jonathan-innis is working on a metrics rewrite -- maybe it's possible to avoid this entirely (or a at a different frequency) somehow.